Structure cleanup: - Moved sources around to create hierarchical maven project structure - Removed dangling .iml file - Removed unirest package-exports from the apps, only needed for the sol-onos-service module
