+++ sol-onos-apps/pom.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>sol-onos</artifactId>
+        <groupId>edu.unc</groupId>
+        <version>1.1</version>
+    </parent>
+
+    <modelVersion>4.0.0</modelVersion>
+    <packaging>pom</packaging>
+
+    <artifactId>sol-onos-apps</artifactId>
+    <version>1.1</version>
+
+    <modules>
+        <module>sol-onos-te-app</module>
+        <module>sol-onos-sc-app</module>
+    </modules>
+
+    <dependencies>
+
+        <dependency>
+            <groupId>edu.unc</groupId>
+            <artifactId>sol-onos-service</artifactId>
+            <version>1.1</version>
+        </dependency>
+
+    </dependencies>
+
+</project>
