+++ sol-onos-service/src/main/java/edu/unc/sol/app/Resource.java
