+++ sol-onos-service/src/main/java/edu/unc/sol/service/TrafficClassDecomposer.java
