+++ sol-onos-service/src/main/resources/features.xml
