--- src/main/resources/solService/css.html
@@ -1,1 +1,0 @@
-<link rel="stylesheet" href="app/view/solService/solService.css">
