--- src/main/resources/solService/js.html
@@ -1,1 +1,0 @@
-<script src="app/view/solService/solService.js"></script>
